+++ .cargo/audit.toml
@@ -1,0 +1,2 @@
+[advisories]
+ignore = ["RUSTSEC-2024-0384", "RUSTSEC-2026-0173", "RUSTSEC-2023-0071", "RUSTSEC-2026-0098", "RUSTSEC-2026-0104", "RUSTSEC-2026-0099", "RUSTSEC-2026-0138", "RUSTSEC-2026-0002"]
